refactor: extract roadmap catalog seed/sync from problems.dao
Moves the roadmap catalog seed data (legacyRoadmapCatalogSeed,
leetcodeRoadmapCatalogSeed, roadmapCatalogSeed, legacyRoadmapFlag,
catalogIdFor) and sync functions (seedLeetCodeStudyPlanProblems,
syncLegacyRoadmapMemberships, syncLeetCodeStudyPlanMemberships,
syncRoadmapCatalog) into a new problems.roadmap.ts file.

problems.dao.ts imports only syncRoadmapCatalog from the new module.
Public problemsDao method set is unchanged. No call site outside
problems.dao.ts was modified.
